+++ R/launchers.R
@@ -1,19 +1,3 @@
-# Copyright (C) 2023-2025 Hibiki AI Limited <info@hibiki-ai.com>
-#
-# This file is part of mirai.
-#
-# mirai is free software: you can redistribute it and/or modify it under the
-# terms of the GNU General Public License as published by the Free Software
-# Foundation, either version 3 of the License, or (at your option) any later
-# version.
-#
-# mirai is distributed in the hope that it will be useful, but WITHOUT ANY
-# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
-# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License along with
-# mirai. If not, see <https://www.gnu.org/licenses/>.
-
 # mirai ------------------------------------------------------------------------
 
 #' Launch Daemon
@@ -66,20 +50,18 @@
 #'
 #' @export
 #'
-launch_local <- function(n = 1L, ..., tls = NULL, .compute = "default") {
-
-  if (missing(.compute)) .compute <- .[["cp"]]
+launch_local <- function(n = 1L, ..., tls = NULL, .compute = NULL) {
+  if (is.null(.compute)) .compute <- .[["cp"]]
   envir <- ..[[.compute]]
   is.null(envir) && stop(._[["daemons_unset"]])
   url <- envir[["urls"]]
-  write_args <- if (length(envir[["msgid"]])) wa3 else wa2
+  write_args <- if (is.null(envir[["msgid"]])) wa2 else wa3
   dots <- if (missing(..1)) envir[["dots"]] else parse_dots(...)
   output <- attr(dots, "output")
   if (is.null(tls)) tls <- envir[["tls"]]
   for (i in seq_len(n))
     launch_daemon(write_args(url, dots, next_stream(envir), tls), output)
   n
-
 }
 
 #' Launch Daemon
@@ -101,9 +83,14 @@ launch_local <- function(n = 1L, ..., tls = NULL, .compute = "default") {
 #' @rdname launch_local
 #' @export
 #'
-launch_remote <- function(n = 1L, remote = remote_config(), ..., tls = NULL, .compute = "default") {
-
-  if (missing(.compute)) .compute <- .[["cp"]]
+launch_remote <- function(
+  n = 1L,
+  remote = remote_config(),
+  ...,
+  tls = NULL,
+  .compute = NULL
+) {
+  if (is.null(.compute)) .compute <- .[["cp"]]
   if (!is.numeric(n) && inherits(n, c("miraiCluster", "miraiNode"))) {
     .compute <- attr(n, "id")
     n <- max(length(n), 1L)
@@ -112,7 +99,7 @@ launch_remote <- function(n = 1L, remote = remote_config(), ..., tls = NULL, .co
   envir <- ..[[.compute]]
   is.null(envir) && stop(._[["daemons_unset"]])
   url <- envir[["urls"]]
-  write_args <- if (length(envir[["msgid"]])) wa3 else wa2
+  write_args <- if (is.null(envir[["msgid"]])) wa2 else wa3
   dots <- if (missing(..1)) envir[["dots"]] else parse_dots(...)
   if (is.null(tls)) tls <- envir[["tls"]]
 
@@ -121,11 +108,9 @@ launch_remote <- function(n = 1L, remote = remote_config(), ..., tls = NULL, .co
   quote <- remote[["quote"]]
 
   if (length(command)) {
-
     args <- remote[["args"]]
 
     if (is.list(args)) {
-
       tunnel <- remote[["tunnel"]]
 
       if (tunnel) {
@@ -138,35 +123,50 @@ launch_remote <- function(n = 1L, remote = remote_config(), ..., tls = NULL, .co
 
       if (length(args) == 1L) {
         args <- args[[1L]]
-
       } else if (n == 1L || n == length(args)) {
-
         cmds <- character(length(args))
         for (i in seq_along(args))
-          cmds[i] <- sprintf("%s -e %s", rscript, write_args(url, dots, next_stream(envir), tls))
+          cmds[i] <- sprintf(
+            "%s -e %s",
+            rscript,
+            write_args(url, dots, next_stream(envir), tls)
+          )
 
         for (i in seq_along(args))
-          system2(command, args = `[<-`(args[[i]], find_dot(args[[i]]), if (quote) shQuote(cmds[i]) else cmds[i]), wait = FALSE)
+          system2(
+            command,
+            args = `[<-`(
+              args[[i]],
+              find_dot(args[[i]]),
+              if (quote) shQuote(cmds[i]) else cmds[i]
+            ),
+            wait = FALSE
+          )
 
         return(`class<-`(cmds, "miraiLaunchCmd"))
-
       } else {
         stop(._[["arglen"]])
       }
-
     }
   }
 
   cmds <- character(n)
   for (i in seq_len(n))
-    cmds[i] <- sprintf("%s -e %s", rscript, write_args(url, dots, next_stream(envir), tls))
+    cmds[i] <- sprintf(
+      "%s -e %s",
+      rscript,
+      write_args(url, dots, next_stream(envir), tls)
+    )
 
   if (length(command))
     for (cmd in cmds)
-      system2(command, args = `[<-`(args, find_dot(args), if (quote) shQuote(cmd) else cmd), wait = FALSE)
+      system2(
+        command,
+        args = `[<-`(args, find_dot(args), if (quote) shQuote(cmd) else cmd),
+        wait = FALSE
+      )
 
   `class<-`(cmds, "miraiLaunchCmd")
-
 }
 
 #' Generic and SSH Remote Launch Configuration
@@ -226,11 +226,20 @@ launch_remote <- function(n = 1L, remote = remote_config(), ..., tls = NULL, .co
 #'
 #' @export
 #'
-remote_config <- function(command = NULL, args = c("", "."), rscript = "Rscript", quote = FALSE) {
-
+remote_config <- function(
+  command = NULL,
+  args = c("", "."),
+  rscript = "Rscript",
+  quote = FALSE
+) {
   if (is.list(args)) lapply(args, find_dot) else find_dot(args)
-  list(command = command, args = args, rscript = rscript, quote = quote, tunnel = FALSE)
-
+  list(
+    command = command,
+    args = args,
+    rscript = rscript,
+    quote = quote,
+    tunnel = FALSE
+  )
 }
 
 #' SSH Remote Launch Configuration
@@ -307,8 +316,13 @@ remote_config <- function(command = NULL, args = c("", "."), rscript = "Rscript"
 #' @rdname remote_config
 #' @export
 #'
-ssh_config <- function(remotes, tunnel = FALSE, timeout = 10, command = "ssh", rscript = "Rscript") {
-
+ssh_config <- function(
+  remotes,
+  tunnel = FALSE,
+  timeout = 10,
+  command = "ssh",
+  rscript = "Rscript"
+) {
   premotes <- lapply(remotes, parse_url)
   hostnames <- lapply(premotes, .subset2, "hostname")
   ports <- lapply(premotes, .subset2, "port")
@@ -316,28 +330,38 @@ ssh_config <- function(remotes, tunnel = FALSE, timeout = 10, command = "ssh", r
   args <- vector(mode = "list", length = length(remotes))
   for (i in seq_along(args)) {
     args[[i]] <- c(
-      sprintf("-o ConnectTimeout=%s -fTp %s", as.character(timeout), ports[[i]]),
+      sprintf(
+        "-o ConnectTimeout=%s -fTp %s",
+        as.character(timeout),
+        ports[[i]]
+      ),
       hostnames[[i]],
       "."
     )
   }
 
-  list(command = command, args = args, rscript = rscript, quote = TRUE, tunnel = isTRUE(tunnel))
-
+  list(
+    command = command,
+    args = args,
+    rscript = rscript,
+    quote = TRUE,
+    tunnel = isTRUE(tunnel)
+  )
 }
 
 #' URL Constructors
 #'
-#' `host_url` constructs a valid host URL (at which daemons may connect)
-#' based on the computer's hostname. This may be supplied directly to the `url`
+#' `host_url()` constructs a valid host URL (at which daemons may connect) based
+#' on the computer's IP address. This may be supplied directly to the `url`
 #' argument of [daemons()].
 #'
-#' `host_url` relies on using the host name of the computer rather than an IP
-#' address and typically works on local networks, although this is not always
-#' guaranteed. If unsuccessful, substitute an IPv4 or IPv6 address in place of
-#' the hostname.
+#' `host_url()` will return a vector of URLs if multiple network adapters are in
+#' use, and each will be named by the interface name (adapter friendly name on
+#' Windows). If this entire vector is passed to the `url` argument of functions
+#' such as `daemons()`, the first URL is used. If no suitable IP addresses are
+#' detected, the computer's hostname will be used as a fallback.
 #'
-#' `local_url` generates a random URL for the platform's default inter-process
+#' `local_url()` generates a random URL for the platform's default inter-process
 #' communications transport: abstract Unix domain sockets on Linux, Unix domain
 #' sockets on MacOS, Solaris and other POSIX platforms, and named pipes on
 #' Windows.
@@ -350,7 +374,8 @@ ssh_config <- function(remotes, tunnel = FALSE, timeout = 10, command = "ssh", r
 #'   connecting from. For `local_url`, is only taken into account if
 #'   `tcp = TRUE`.
 #'
-#' @return A character string comprising a valid URL.
+#' @return A character vector (comprising a valid URL or URLs), named for
+#'   `host_url()`.
 #'
 #' @examples
 #' host_url()
@@ -359,15 +384,22 @@ ssh_config <- function(remotes, tunnel = FALSE, timeout = 10, command = "ssh", r
 #'
 #' @export
 #'
-host_url <- function(tls = FALSE, port = 0)
-  sprintf(
-    "%s://%s:%s",
-    if (tls) "tls+tcp" else "tcp", Sys.info()[["nodename"]], as.character(port)
+host_url <- function(tls = FALSE, port = 0) {
+  ip <- ip_addr()
+  `names<-`(
+    sprintf(
+      "%s://%s:%d",
+      if (tls) "tls+tcp" else "tcp",
+      if (any(nzchar(ip))) ip else Sys.info()[["nodename"]],
+      as.integer(port)
+    ),
+    names(ip)
   )
+}
 
 #' URL Constructors
 #'
-#' `local_url` constructs a URL suitable for local daemons, or for use with
+#' `local_url()` constructs a URL suitable for local daemons, or for use with
 #' SSH tunnelling. This may be supplied directly to the `url` argument of
 #' [daemons()].
 #'
@@ -383,18 +415,14 @@ host_url <- function(tls = FALSE, port = 0)
 #' @export
 #'
 local_url <- function(tcp = FALSE, port = 0)
-  if (tcp)
-    sprintf("tcp://127.0.0.1:%d", as.integer(port)) else
-      sprintf("%s%s", .urlscheme, random(12L))
+  if (tcp) sprintf("tcp://127.0.0.1:%d", as.integer(port)) else
+    sprintf("%s%s", .urlscheme, random(12L))
 
 #' @export
 #'
 print.miraiLaunchCmd <- function(x, ...) {
-
-  for (i in seq_along(x))
-    cat(sprintf("[%d]\n%s\n\n", i, x[i]), file = stdout())
+  for (i in seq_along(x)) cat(sprintf("[%d]\n%s\n\n", i, x[i]), file = stdout())
   invisible(x)
-
 }
 
 # internals --------------------------------------------------------------------