+++ R/promises.R
@@ -1,19 +1,3 @@
-# Copyright (C) 2023-2025 Hibiki AI Limited <info@hibiki-ai.com>
-#
-# This file is part of mirai.
-#
-# mirai is free software: you can redistribute it and/or modify it under the
-# terms of the GNU General Public License as published by the Free Software
-# Foundation, either version 3 of the License, or (at your option) any later
-# version.
-#
-# mirai is distributed in the hope that it will be useful, but WITHOUT ANY
-# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
-# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License along with
-# mirai. If not, see <https://www.gnu.org/licenses/>.
-
 # mirai.promises ---------------------------------------------------------------
 
 # Joe Cheng (3 Apr 2024):
@@ -59,11 +43,9 @@
 #' @exportS3Method promises::as.promise
 #'
 as.promise.mirai <- function(x) {
-
   promise <- .subset2(x, "promise")
 
   if (is.null(promise)) {
-
     promise <- if (unresolved(x)) {
       promises::promise(
         function(resolve, reject) .keep(x, environment())
@@ -87,11 +69,9 @@ as.promise.mirai <- function(x) {
     }
 
     `[[<-`(x, "promise", promise)
-
   }
 
   promise
-
 }
 
 #' Make mirai_map Promise
@@ -129,14 +109,12 @@ as.promise.mirai <- function(x) {
 #' @exportS3Method promises::as.promise
 #'
 as.promise.mirai_map <- function(x) {
-
   promise <- attr(x, "promise")
 
   if (is.null(promise))
     attr(x, "promise") <- promise <- promises::promise_all(.list = x)
 
   promise
-
 }
 
 #' @exportS3Method promises::is.promising